Add files via upload
This is a fully functional version of 9/6 Jacks or Better. (It's called "9/6" because the Full House pays 9 times your bet and the Flush pays 6 times your bet). Just type the cards you want to keep, all as one big number. So if you want to keep the pair of fives here: 2♦ 5♥ T♥ 5♠ J♣. You would type in "24" as you want to keep the second and fourth card.